LPS-127508 Move common files to headless-delivery-api
